requirements.txt
invisible-watermark
accelerate #似乎会和transformer-engine冲突,安装accelerate之后如果报错可以pip uninstall transformer-engine
diffusers==0.20.2
ipython==8.14.0
matplotlib==3.7.2
numpy==1.25.1
opencv_python==4.6.0.66
Pillow==9.5.0
pytorch_lightning==2.0.1
seaborn==0.12.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
pyyaml
hydra-core==1.3.2
-transformer-engine